clima/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.17
django-bootstrap-v5==1.0.11
djangorestframework==3.15.2
filelock==3.17.0
gunicorn==23.0.0
idna==3.10
lxml==5.3.1
packaging==24.2
pillow==11.0.0
PySocks==1.7.1
requests==2.32.3
snscrape==0.7.0.20230622
soupsieve==2.6
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
